requirements.txt
asgiref==3.7.2
Django==4.2.15
django-filter==23.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
drf-yasg==1.21.7
et_xmlfile==2.0.0
inflection==0.5.1
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
django-cors-headers==4.4.0
